requirements.txt
pydic==1.0
u-msgpack-python==1.6